move-only: split CNode into new node/connection.h module
Move per-peer code from net into a new node/connection module. This
is intended to capture logic that is lower-level than the p2p protocol
(done in net_processing), but higher-level than raw bits-over-the-wire
(handled by Sock).
