+++ src/Makefile.am
@@ -218,6 +218,7 @@ BITCOIN_CORE_H = \
   node/chainstatemanager_args.h \
   node/coin.h \
   node/coins_view_args.h \
+  node/connection.h \
   node/connection_types.h \
   node/context.h \
   node/database_args.h \
@@ -417,6 +418,7 @@ libbitcoin_node_a_SOURCES = \
   node/chainstatemanager_args.cpp \
   node/coin.cpp \
   node/coins_view_args.cpp \
+  node/connection.cpp \
   node/connection_types.cpp \
   node/context.cpp \
   node/database_args.cpp \
